+++ eagle/eagle/EaglePlanJob.c
@@ -9,7 +9,7 @@ EaglePlanJob* EaglePlanJob_New(EaglePlan *plan, int allocatedBuffers)
     
     // initialize all buffers now
     job->allocatedBuffers = allocatedBuffers;
-    job->buffers = (EaglePage**) calloc(job->allocatedBuffers, sizeof(EaglePage));
+    job->buffers = (EaglePage**) calloc((size_t) job->allocatedBuffers, sizeof(EaglePage));
     for(int i = 0; i < job->allocatedBuffers; ++i) {
         job->buffers[i] = EaglePage_Alloc(plan->pageSize);
     }
